requirements.txt
beautifulsoup4
requests
django-environ

